chore(package) rename package to swent.skillswap
Warmup
