+++ app/build.gradle.kts
@@ -7,11 +7,11 @@ plugins {
 }
 
 android {
-    namespace = "com.android.sample"
+    namespace = "com.swent.skillswap"
     compileSdk = 34
 
     defaultConfig {
-        applicationId = "com.android.sample"
+        applicationId = "com.swent.skillswap"
         minSdk = 28
         targetSdk = 34
         versionCode = 1
