+++ app/src/androidTest/java/com/swent/skillswap/ExampleInstrumentedTest.kt
@@ -1,10 +1,11 @@
-package com.android.sample
+package com.swent.skillswap
 
 import androidx.compose.ui.test.junit4.createAndroidComposeRule
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import com.android.sample.screen.SecondScreen
+import com.swent.skillswap.screen.MainScreen
 import com.kaspersky.kaspresso.testcases.api.testcase.TestCase
 import io.github.kakaocup.compose.node.element.ComposeScreen
+import org.junit.Assert.*
 import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
@@ -15,17 +16,17 @@ import org.junit.runner.RunWith
  * See [testing documentation](http://d.android.com/tools/testing).
  */
 @RunWith(AndroidJUnit4::class)
-class SecondActivityTest : TestCase() {
+class MainActivityTest : TestCase() {
 
-  @get:Rule val composeTestRule = createAndroidComposeRule<SecondActivity>()
+  @get:Rule val composeTestRule = createAndroidComposeRule<MainActivity>()
 
   @Test
   fun test() = run {
-    step("Start Second Activity") {
-      ComposeScreen.onComposeScreen<SecondScreen>(composeTestRule) {
+    step("Start Main Activity") {
+      ComposeScreen.onComposeScreen<MainScreen>(composeTestRule) {
         simpleText {
           assertIsDisplayed()
-          assertTextEquals("Hello Robolectric!")
+          assertTextEquals("Hello Android!")
         }
       }
     }
