+++ app/src/main/java/com/swent/skillswap/resources/C.kt
@@ -1,4 +1,4 @@
-package com.android.sample.resources
+package com.swent.skillswap.resources
 
 // Like R, but C
 object C {
