+++ app/src/main/java/com/swent/skillswap/ui/theme/Color.kt
@@ -1,4 +1,4 @@
-package com.android.sample.ui.theme
+package com.swent.skillswap.ui.theme
 
 import androidx.compose.ui.graphics.Color
