+++ app/src/main/java/com/swent/skillswap/ui/theme/Theme.kt
@@ -1,4 +1,4 @@
-package com.android.sample.ui.theme
+package com.swent.skillswap.ui.theme
 
 import android.app.Activity
 import android.os.Build
