+++ app/src/test/java/com/swent/skillswap/PointTest.kt
@@ -1,4 +1,4 @@
-package com.android.sample
+package com.swent.skillswap
 
 import org.junit.Assert.assertEquals
 import org.junit.Test
